Ok... What all to list...
First off... there's some bugs with the code that need to be fixed. 

1) When you exit the app, it crashes the ps3... probably me not throwing a callback to GL2D somewhere

2) This ONLY runs with the pad code... seriously comment the pad code out from main and it won't run. I have NO clue why.

With that out of the way, my list of changes:

/* GL2D */

Moved allImage to GL2D.cpp since it's the only thing that uses it.

Overloaded drawSprite() to accept SpriteA class

/* Sprite */

made setCoord() virtual so that SpriteA could use it's png vector to calculate

/* SpriteA */

This class holds multiple PNG images and provides a method of animating them. A lot is meant to be added onto this class once the above listed bugs are first handled. Currently animation fps is not implemented, pattern recognition for loading files is not implemented, and the ability to load a sprite sheet is not implemented. 

The constructor takes 3 arguments:
the partial path: /USRDIR/Sprites/bla
the first number of image to load: /USRDIR/Sprites/bla1.png so 1
the last number of image to load: /USRDIR/Sprites/bla25.png so 25
The constructor will then load up all of the images within the range.

There are functions to start and stop the animation with the option of designating a certain frame one which to start/stop on.

increment() increments the animation and loops to the beginning when the end is reached (note currently, I staticly designated 39. This is because of debugging purposes. Change this to frame.size() to test your own images.